split getFileMode() by platform
