+++ ziti/cmd/unwrap_identity.go
@@ -7,9 +7,7 @@ import (
 	"github.com/spf13/cobra"
 	"io"
 	"os"
-	"runtime"
 	"strings"
-	"syscall"
 )
 
 type IdentityConfigFile struct {
@@ -94,25 +92,4 @@ func NewUnwrapIdentityFileCommand(out io.Writer, errOut io.Writer) *cobra.Comman
 	cmd.Flags().StringVarP(&outCaFile, "ca", "", "", "output ca bundle file, defaults to ./<root>.ca")
 
 	return cmd
-}
-
-func getFileMode(isPrivateKey bool) os.FileMode {
-	// Default modes before umask:
-	// - Private keys: 0600 (rw-------)
-	// - Public files: 0644 (rw-r--r--)
-	mode := os.FileMode(0644)
-	if isPrivateKey {
-		mode = os.FileMode(0600)
-	}
-
-	if runtime.GOOS != "windows" {
-		// Get current umask
-		oldMask := syscall.Umask(0)
-		syscall.Umask(oldMask) // Restore original umask
-
-		// Apply umask to our default mode
-		mode &= ^os.FileMode(oldMask)
-	}
-
-	return mode
 }
